video-generation/backend/requirements.txt
celery==5.5.1
redis==5.2.1
python-dotenv==1.1.0
qdrant-client==1.14.1
openai==1.75.0
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-oauthlib==1.2.2
requests==2.32.3
djangorestframework==3.16.0
django-cors-headers==3.14.0
Django==5.2
django-celery-beat==2.8.0
yt-dlp==2025.3.31
djangorestframework-simplejwt==5.5.0
cryptography==44.0.2
elevenlabs==1.56.0
replicate==1.0.4
aiohttp==3.11.18
psycopg2-binary==2.9.10
gunicorn==23.0.0



